Moved LogStartTestInfo into ListenerBase
LogStartTestInfo only needed by listeners; method was accessible by tests before. Formatted spacing
