+++ src/test/java/autoFramework/ListenerBase.java
@@ -9,17 +9,20 @@ import pages.Pages;
 
 import java.io.File;
 import java.io.IOException;
+import java.util.ArrayList;
+import java.util.List;
 
 /** Extended by listener classes. */
 public class ListenerBase extends AutoLogger
 {
-    // TODO: should move these Strings and getter/ setters to AutoLogger
     private static String testName;
     private static String packageName;
     private static String className;
     private static String screenshotDir;
+    private static String testLevel;
     private static final ExtentFactory extentFactory = new ExtentFactory();
     private final ExtentReportManager extentReportManager = new ExtentReportManager();
+    private final TestExecutionContext testExecutionContext = new TestExecutionContext();
 
     // Getter methods for Strings
     public String GetTestName() { return testName; }
@@ -39,10 +42,7 @@ public class ListenerBase extends AutoLogger
 
     /** Splits PackageName.ClassName and sets them.
      * @param name PackageName.ClassName*/
-    public void SetPackageClassName(String name)
-    {
-        SplitPackageClassName(name);
-    }
+    public void SetPackageClassName(String name) { SplitPackageClassName(name); }
 
     private void SplitPackageClassName(String packageClassName)
     {
@@ -54,28 +54,18 @@ public class ListenerBase extends AutoLogger
 
     // Methods for Extent Report
     /** Set up ExtentReport and ExtentSparkReporter. */
-    public ExtentReports SetUpExtentReporter() {
-        return extentReportManager.GetInstance();
-    }
+    public ExtentReports SetUpExtentReporter() { return extentReportManager.GetInstance(); }
 
     // TODO: Keep, may not be needed
-    public ExtentFactory GetExtentFactory() {
-        return extentFactory;
-    }
+    public ExtentFactory GetExtentFactory() { return extentFactory; }
 
-    public static ExtentTest GetExtentTest() {
-        return extentFactory.GetExtentTest();
-    }
+    public static ExtentTest GetExtentTest() { return extentFactory.GetExtentTest(); }
 
     /** Set up ExtentTest at the start of test*/
-    public void SetExtentTest(ExtentTest extentTestObject) {
-        extentFactory.SetExtentTest(extentTestObject);
-    }
+    public void SetExtentTest(ExtentTest extentTestObject) { extentFactory.SetExtentTest(extentTestObject); }
 
     /** Removes ExtentTest from thread. */
-    public void RemoveExtentTest() {
-        extentFactory.RemoveExtentObject();
-    }
+    public void RemoveExtentTest() { extentFactory.RemoveExtentObject(); }
 
     // Screenshot methods
     // TODO: Screenshot first failure, last failure, all failures?
@@ -100,8 +90,34 @@ public class ListenerBase extends AutoLogger
         screenshotDir = "./failedTests/" + GetPackageClassName() + "/" + GetTestEndDate() + GetTestName() + ".png";
     }
 
-    public String GetScreenShotDir()
-    {
-        return screenshotDir;
+    public String GetScreenShotDir() { return screenshotDir; }
+
+    /** Logs testExecutionContext to provide information at the start of the test. */
+    public void LogStartTestInfo(String testName, String packageClassName) throws ClassNotFoundException {
+
+        testExecutionContext.GetTestInfoContext(testName, packageClassName);
+
+        List<String> messages = new ArrayList<>();
+        messages.add("");
+        messages.add("==========================================================================");
+        messages.add("     Starting test  : " + testName);
+        messages.add("     Test package   : " + testExecutionContext.getPackageName());
+        messages.add("     Test class     : " + testExecutionContext.getClassName());
+        messages.add("     Description    : " + testExecutionContext.getDescription());
+        messages.add("     Test level     : " + testExecutionContext.getLevel());
+        //messages.add("     Test categories:  " + testExecutionContext.getCategories());
+        messages.add("==========================================================================");
+
+        // set test level to be categorized by Extent Report
+        // need to save level in static var to be used in other methods
+        testLevel = testExecutionContext.getLevel();
+
+        // Log test info
+        for(String message : messages)
+        {
+            SetUpInfo(message);
+        }
     }
+
+    public static String GetTestLevel() { return testLevel; }
 }
